refactor serialization / recursion limit handling
