+++ manticore/core/smtlib/__init__.py
@@ -3,7 +3,6 @@ from .constraints import ConstraintSet  # noqa
 from .solver import *  # noqa
 from . import operators as Operators  # noqa
 
-
 import logging
 logger = logging.getLogger(__name__)
