+++ manticore/core/state.py
@@ -1,7 +1,5 @@
 import copy
 import logging
-import pickle
-import sys
 
 from .smtlib import solver, Bool
 from ..utils.helpers import issymbolic
@@ -429,41 +427,3 @@ class State(Eventful):
         :param str message: Longer description
         """
         self._publish('will_generate_testcase', name, message)
-
-
-class StateSerializer(object):
-    """
-    StateSerializer can serialize and deserialize :class:`~manticore.core.state.State` objects from and to
-    stream-like objects.
-    """
-
-    def __init__(self):
-        pass
-
-    def serialize(self, state, f):
-        raise NotImplementedError
-
-    def deserialize(self, f):
-        raise NotImplementedError
-
-
-class PickleSerializer(StateSerializer):
-    DEFAULT_RECURSION: int = 0x100000  # 1M
-    MAX_RECURSION: int = 0x1000000  # 16.7M
-
-    def __init__(self):
-        super().__init__()
-        sys.setrecursionlimit(PickleSerializer.DEFAULT_RECURSION)
-
-    def serialize(self, state, f):
-        try:
-            f.write(pickle.dumps(state, 2))
-        except RuntimeError:
-            if sys.getrecursionlimit() >= PickleSerializer.MAX_RECURSION:
-                raise Exception(f'PickleSerializer recursion limit surpassed {PickleSerializer.MAX_RECURSION}, aborting')
-            logger.info(f'Recursion maximum {sys.getrecursionlimit()} hit, increasing')
-            sys.setrecursionlimit(sys.getrecursionlimit() + PickleSerializer.DEFAULT_RECURSION)
-            self.serialize(state, f)
-
-    def deserialize(self, f):
-        return pickle.load(f)
